Completed the excel loading part of the app
User can upload an excel file and the contents of the excel file will be inserted into a database using AJAX request and a Django view. If it is successful, an alert box indicating success is shown. Else, an alert box indicating error is shown. While the ajax request is loading, a loading gif tells the user that the excel file is being processed.