Fixes issue #103, adds fix for newlines in assignment comments
We implemented a change that should close issue #103, in which
newline characters in assignment comments were being rendered in
the HTML rather than being converted into newlines. We had to
import the Python regular expression library in order to change
all instances of the newline character in grader comments to be a
<br /> tag, which will actually be rendered as a new line in HTML.
We have ran our fix against all of the project's tests, and the
software still passes everything. We also included documentation
of our changes via comments in the code.