Modify fix to comply with XHTML, sign comments
In this commit, we altered our fix slightly so that it will allow
the rendered web pages to conform to the XHTML standards. We were
unable to confirm for certain whether or not the entire project is
intended to comply with these standards, but thought it would not
hurt to do so for our changes in the case that that is required.
We also signed all of the comments so that it is easier to find
the changes we made.