Replace newline literals in grader comments with break tags (#103) (#213)
* Fixes issue #103, adds fix for newlines in assignment comments

We implemented a change that should close issue #103, in which
newline characters in assignment comments were being rendered in
the HTML rather than being converted into newlines. We had to
import the Python regular expression library in order to change
all instances of the newline character in grader comments to be a
<br /> tag, which will actually be rendered as a new line in HTML.
We have ran our fix against all of the project's tests, and the
software still passes everything. We also included documentation
of our changes via comments in the code.

* Modify fix to comply with XHTML, sign comments

In this commit, we altered our fix slightly so that it will allow
the rendered web pages to conform to the XHTML standards. We were
unable to confirm for certain whether or not the entire project is
intended to comply with these standards, but thought it would not
hurt to do so for our changes in the case that that is required.
We also signed all of the comments so that it is easier to find
the changes we made.

* Fix issue raised by initial fix

Our initial fix made a comparison to null that apparently is not
acceptable based on project standards. We changed it so that it
should conform to the expectations of the project.

* Add another change to the initial fix

We originally returned nothing when the grader comment is the
empty string in the function we edited, but really just want to
return the unedited comment.

* Fix comments

Per request, I have removed our usernames from the comments in our
code. No other changes.

* Modify test data to include instance of newline literals in grader comment

* Tweak and fix original implementation

* Add formatted_grader_comment test

* Change quotes around repl string to double

* Tweak null handling

* Correct null-catching condition

* Add explicit NULL for grader_comment in test data

* Remove extra newline from new test

Co-authored-by: harryrs <harryrs@umich.edu>